requirements/requirements-dev.txt
autopep8==2.0.1
pycodestyle==2.10.0
django_debug_toolbar==3.8.1